+++ packages/game-art/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @bombfarm/game-art
 
+## 0.6.3
+
+### Patch Changes
+
+- Updated dependencies [fdc8b8e]
+  - @bombfarm/domain@1.3.1
+
 ## 0.6.2
 
 ### Patch Changes